size_t handling for 32-bit platforms

+++ GraphBLAS/Source/GB_realloc_memory.c
@@ -84,7 +84,8 @@ void *GB_realloc_memory     // pointer to reallocated block of memory, or
     (*ok) = GB_size_t_multiply (&newsize, nitems_new, size_of_item)
          && GB_size_t_multiply (&oldsize, nitems_old, size_of_item) ;
 
-    if (!(*ok) || nitems_new > GB_NMAX || size_of_item > GB_NMAX)
+    if (!(*ok) || (((uint64_t) nitems_new) > GB_NMAX)
+               || (((uint64_t) size_of_item) > GB_NMAX))
     { 
         // overflow
         (*ok) = false ;
